stile nuovo

+++ database/migrations/2025_05_09_100909_create_articles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('articles', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->float('price');
+            $table->decimal('price', 6, 2);
             $table->string('image');
             $table->unsignedBigInteger('user_id')->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('SET NULL');
